Direct API move card details out of additionalData
Not sure if the documentation has chaned or whether this has always
been wrong, but should be fixed now. Note that direct CC payments
are disabled as an account feature by default (even on test) and
must be requested from Adyen before it will work. The "Not allowed"
error message will be returned otherwise.